CRT-33 refactor: 레이어 구조에 따라 패키지 구조 변경
