+++ src/main/java/knu/ridi/charactalk/character/repository/CharacterRepository.java
@@ -1,5 +1,6 @@
-package knu.ridi.charactalk.character;
+package knu.ridi.charactalk.character.repository;
 
+import knu.ridi.charactalk.character.domain.Character;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CharacterRepository extends JpaRepository<Character, Long> {
