+++ src/main/java/knu/ridi/charactalk/chat/domain/SenderType.java
@@ -1,4 +1,4 @@
-package knu.ridi.charactalk.chat;
+package knu.ridi.charactalk.chat.domain;
 
 public enum SenderType {
     MEMBER, CHARACTER
