+++ src/main/java/knu/ridi/charactalk/chat/repository/ChatRepository.java
@@ -1,5 +1,6 @@
-package knu.ridi.charactalk.chat;
+package knu.ridi.charactalk.chat.repository;
 
+import knu.ridi.charactalk.chat.domain.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRepository extends JpaRepository<Chat, Long> {
