+++ src/main/java/knu/ridi/charactalk/chat/supporter/ChatStreamManager.java
@@ -1,6 +1,6 @@
 package knu.ridi.charactalk.chat.supporter;
 
-import knu.ridi.charactalk.chat.dto.ChatStreamResponse;
+import knu.ridi.charactalk.chat.api.dto.ChatStreamResponse;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Sinks;
@@ -11,11 +11,12 @@ import java.util.concurrent.ConcurrentHashMap;
 
 @Component
 public class ChatStreamManager {
+
     private final Map<Long, Sinks.Many<ChatStreamResponse>> sinks = new ConcurrentHashMap<>();
 
     public Flux<ChatStreamResponse> subscribe(Long memberId) {
         return sinks.computeIfAbsent(memberId,
-                k -> Sinks.many().multicast().onBackpressureBuffer()
+            k -> Sinks.many().multicast().onBackpressureBuffer()
         ).asFlux();
     }
