+++ src/main/java/knu/ridi/charactalk/chatroom/domain/ChatRoomType.java
@@ -1,4 +1,4 @@
-package knu.ridi.charactalk.chatroom;
+package knu.ridi.charactalk.chatroom.domain;
 
 public enum ChatRoomType {
     SINGLE, GROUP
