+++ src/main/java/knu/ridi/charactalk/chatroom/repository/ChatRoomRepository.java
@@ -1,5 +1,6 @@
-package knu.ridi.charactalk.chatroom;
+package knu.ridi.charactalk.chatroom.repository;
 
+import knu.ridi.charactalk.chatroom.domain.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
