+++ src/main/java/knu/ridi/charactalk/member/domain/Gender.java
@@ -1,0 +1,6 @@
+package knu.ridi.charactalk.member.domain;
+
+public enum Gender {
+
+    MALE, FEMALE
+}
