--- src/main/java/knu/ridi/charactalk/member/entity/Gender.java
@@ -1,6 +1,0 @@
-package knu.ridi.charactalk.member.entity;
-
-public enum Gender {
-
-    MALE, FEMALE
-}
