+++ src/main/java/knu/ridi/charactalk/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package knu.ridi.charactalk.member;
+package knu.ridi.charactalk.member.repository;
 
-import knu.ridi.charactalk.member.entity.Member;
+import knu.ridi.charactalk.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
